Code cleanup in activities slice
